pyproject.toml
[project]
name = "vp-recipe-agent"
version = "0.1.0"
description = "VP Recipe Agent — LLM Evals Homework"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "python-dotenv>=1.0.0",
]
